+++ src/QuikSharp/ISecurity.cs
@@ -1,15 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
 namespace QuikSharp {
+    /// <summary>
+    /// 
+    /// </summary>
     public interface ISecurity {
+        /// <summary>
+        /// 
+        /// </summary>
         [JsonProperty("class_code")]
         string ClassCode { get; set; }
 
+        /// <summary>
+        /// 
+        /// </summary>
         [JsonProperty("sec_code")]
         string SecCode { get; set; }
     }