+++ src/QuikSharp/LuaException.cs
@@ -5,8 +5,24 @@ namespace QuikSharp {
     /// An exception caught on Lua side with a message from Lua
     /// </summary>
     public class LuaException : Exception {
-        public LuaException(string message) : base(message) {
-            
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="message"></param>
+        public LuaException(string message)
+            : base(message) {
+
         }
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    public class TransactionException : LuaException {
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="message"></param>
+        public TransactionException(string message) : base(message) { }
     }
 }