+++ tests/Profiler/Program.cs
@@ -5,18 +5,20 @@ using QuikSharp;
 
 namespace Profiler {
     public class Program {
-        public static void Main() {
-            var qc = new DebugFunctions(Quik.DefaultPort);
+
+        public static void Ping() {
+            var _df = new DebugFunctions(Quik.DefaultPort);
+
             var sw = new Stopwatch();
             Console.WriteLine("Started");
             for (int round = 0; round < 10; round++) {
                 sw.Reset();
                 sw.Start();
-                
+
                 var count = 10000;
                 var array = new Task<string>[count];
                 for (int i = 0; i < array.Length; i++) {
-                    array[i] = qc.Ping();
+                    array[i] = _df.Ping();
                 }
                 for (int i = 0; i < array.Length; i++) {
                     var pong = array[i].Result;
@@ -33,6 +35,40 @@ namespace Profiler {
             }
             Console.WriteLine("Finished");
             Console.ReadKey();
+        }
+
+
+        public static void EchoTransaction() {
+            var _df = new DebugFunctions(Quik.DefaultPort);
+
+            var sw = new Stopwatch();
+            Console.WriteLine("Started");
+            for (int round = 0; round < 10; round++) {
+                sw.Reset();
+                sw.Start();
+
+                var count = 10000;
+                var t = new TransactionSpecification();
+
+                var array = new Task<TransactionSpecification>[count];
+                for (int i = 0; i < array.Length; i++) {
+                    array[i] = _df.Echo(t);
+                }
+                for (int i = 0; i < array.Length; i++) {
+                    var res = array[i].Result;
+                    array[i] = null;
+                }
+
+                sw.Stop();
+                Console.WriteLine("MultiPing takes msecs: " + sw.ElapsedMilliseconds);
+            }
+            Console.WriteLine("Finished");
+            Console.ReadKey();
+        }
+
+        public static void Main() {
+            //Ping();
+            EchoTransaction();
         }
     }
 }